Move the URL sanitization to its own module
The rule does not belong to the tab, which only happens to be one of its
callers, so it moves to a module of its own, with the tab keeping the
sanitizer type exported for compatibility.

The tests of the page context move to the test of the module declaring
it. The ones covering the absence of a browser stay apart, as the test
environment cannot remove the window from a page.
