+++ src/tab.ts
@@ -1,25 +1,13 @@
 import type {EventListener} from './eventManager';
 import {SynchronousEventManager} from './eventManager';
+import type {UrlSanitizer} from './url';
+import {sanitizeUrl} from './url';
+
+export type {UrlSanitizer};
 
 export type TabEvent<T = Record<string, unknown>> = CustomEvent<{tab: Tab} & T>;
 export type TabVisibilityChangeEvent = TabEvent<{visible: boolean}>;
 export type TabUrlChangeEvent = TabEvent<{url: string}>;
-
-export type UrlSanitizer = (url: string) => URL;
-
-/**
- * Normalizes the URL and sanitizes it, if a sanitizer is given.
- *
- * @param url The URL to sanitize.
- * @param sanitizer The sanitizer to apply to the normalized URL.
- *
- * @returns The sanitized URL.
- */
-export function sanitizeUrl(url: string, sanitizer?: UrlSanitizer): string {
-    const normalizedUrl = encodeURI(decodeURI(url));
-
-    return sanitizer !== undefined ? sanitizer(normalizedUrl).toString() : normalizedUrl;
-}
 
 type TabEventMap = {
     focus: TabEvent,
